+++ examples/vello_editor/src/text.rs
@@ -3,7 +3,6 @@
 
 use accesskit::{Node, TreeUpdate};
 use core::default::Default;
-use parley::{editor::SplitString, outputs::PositionedLayoutItem, GenericFamily, StyleProperty};
 use std::time::{Duration, Instant};
 use vello::{
     kurbo::{Affine, Line, Stroke},
@@ -16,8 +15,9 @@ use winit::{
     keyboard::{Key, NamedKey},
 };
 
-pub use parley::outputs::editor::Generation;
-use parley::{FontContext, LayoutContext, PlainEditor, PlainEditorDriver};
+use parley::inputs::{FontContext, GenericFamily, LayoutContext, StyleProperty};
+use parley::outputs::editor::{Generation, PlainEditor, PlainEditorDriver, SplitString};
+use parley::outputs::PositionedLayoutItem;
 
 use crate::access_ids::next_node_id;
