+++ parley/src/tests/test_basic.rs
@@ -1,7 +1,8 @@
 // Copyright 2024 the Parley Authors
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-use crate::{testenv, Alignment, InlineBox};
+use crate::outputs::Alignment;
+use crate::{testenv, InlineBox};
 
 #[test]
 fn plain_multiline_text() {
