+++ parley/src/tests/utils/env.rs
@@ -1,11 +1,13 @@
 // Copyright 2024 the Parley Authors
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-use crate::tests::utils::renderer::{render_layout, ColorBrush, RenderingConfig};
-use crate::{
-    FontContext, FontFamily, FontStack, Layout, LayoutContext, PlainEditor, PlainEditorDriver,
-    RangedBuilder, Rect, StyleProperty,
+use crate::inputs::{
+    FontContext, FontFamily, FontStack, LayoutContext, RangedBuilder, StyleProperty,
 };
+use crate::outputs::editor::{PlainEditor, PlainEditorDriver};
+use crate::outputs::Layout;
+use crate::tests::utils::renderer::{render_layout, ColorBrush, RenderingConfig};
+use crate::Rect;
 use fontique::{Collection, CollectionOptions};
 use std::path::{Path, PathBuf};
 use tiny_skia::{Color, Pixmap};
